requirements.txt
torch
torchvision
torchaudio
triton==3.0.0
transformers==4.49.0
accelerate==0.34.2
huggingface-hub==0.29.1
safetensors==0.4.5
tokenizers==0.21.0
clip
ftfy==6.2.3
regex==2024.9.11
tqdm==4.67.1
qwen-vl-utils==0.0.10
#faiss-gpu 
numpy==1.26.3
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
Pillow==10.2.0
opencv-python==4.10.0.84
matplotlib==3.9.2
seaborn==0.13.2
jsonschema>=4.21.0
requests==2.32.3
